res fixed some warnings reportedly emitted by GCC in winprintf.cpp. res added a method to csCursorConverter to generate a 1bpp cursor,   but foreground and background color are automatically determined   (instead of having to be provided by the user). res fixed VS8 compiler warnings caused by a wrong include order in   tritri.cpp and also made the FABS macro use fabsf().
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20989 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgeom/tritri.cpp
@@ -25,14 +25,14 @@
  * 2005-04-15 ported to CS by Andrew Dai
  */
 
-#include <math.h>
 #include "cssysdef.h"
+#include <math.h>
 
 #include "csgeom/math3d.h"
 #include "csgeom/vector3.h"
 #include "csgeom/segment.h"
 
-#define FABS(x) ((float)fabs(x))        /* implement as is fastest on your machine */
+#define FABS(x) (fabsf(x))        /* implement as is fastest on your machine */
 
 /* if USE_EPSILON_TEST is true then we do a check: 
          if |dv|<SMALL_EPSILON then dv=0.0;
